requirements.txt
Flask~=2.2.2
flask-cors~=3.0.10
Flask-SQLAlchemy~=3.0.2
python-dotenv~=0.21.0
Werkzeug~=2.2.2
psycopg2-binary
